@@ -25,17 +25,20 @@ import java.io.IOException;
 import java.io.PrintStream;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Timeout;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 @RunWith(JUnit4.class)
+@Timeout(value = 10, unit = TimeUnit.MINUTES)
 public class CustomHostnameInstanceIT {
-
   private static final String PROJECT_ID = System.getenv("GOOGLE_CLOUD_PROJECT");
   private static String INSTANCE_NAME;
   private static String ZONE;
@@ -49,8 +52,9 @@ public class CustomHostnameInstanceIT {
         .that(System.getenv(envVarName)).isNotEmpty();
   }
 
-  @BeforeClass
-  public static void setup() throws IOException, ExecutionException, InterruptedException {
+  @BeforeAll
+  public static void setup()
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     final PrintStream out = System.out;
     ByteArrayOutputStream stdOut = new ByteArrayOutputStream();
     System.setOut(new PrintStream(stdOut));
@@ -70,8 +74,9 @@ public class CustomHostnameInstanceIT {
     System.setOut(out);
   }
 
-  @AfterClass
-  public static void cleanUp() throws IOException, ExecutionException, InterruptedException {
+  @AfterAll
+  public static void cleanUp()
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     final PrintStream out = System.out;
     ByteArrayOutputStream stdOut = new ByteArrayOutputStream();
     System.setOut(new PrintStream(stdOut));
@@ -80,13 +85,13 @@ public class CustomHostnameInstanceIT {
     System.setOut(out);
   }
 
-  @Before
+  @BeforeEach
   public void beforeEach() {
     stdOut = new ByteArrayOutputStream();
     System.setOut(new PrintStream(stdOut));
   }
 
-  @After
+  @AfterEach
   public void afterEach() {
     stdOut = null;
     System.setOut(null);
