@@ -16,10 +16,7 @@
 
 package com.example.cloudsql;
 
-import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.logging.Logger;
 import javax.servlet.ServletContextEvent;
@@ -40,94 +37,6 @@ public class ConnectionPoolContextListener implements ServletContextListener {
   private static final String DB_PASS = System.getenv("DB_PASS");
   private static final String DB_NAME = System.getenv("DB_NAME");
 
-  private DataSource createConnectionPool() {
-    // [START cloud_sql_sqlserver_servlet_create]
-    // Note: For Java users, the Cloud SQL JDBC Socket Factory can provide authenticated connections
-    // which is preferred to using the Cloud SQL Proxy with Unix sockets.
-    // See https://github.com/GoogleCloudPlatform/cloud-sql-jdbc-socket-factory for details.
-
-    // The configuration object specifies behaviors for the connection pool.
-    HikariConfig config = new HikariConfig();
-
-    // The following is equivalent to setting the config options below:
-    // jdbc:sqlserver://;user=<DB_USER>;password=<DB_PASS>;databaseName=<DB_NAME>;
-    // socketFactoryClass=com.google.cloud.sql.sqlserver.SocketFactory;
-    // socketFactoryConstructorArg=<INSTANCE_CONNECTION_NAME>
-    
-    // See the link below for more info on building a JDBC URL for the Cloud SQL JDBC Socket Factory
-    // https://github.com/GoogleCloudPlatform/cloud-sql-jdbc-socket-factory#creating-the-jdbc-url
-
-    // Configure which instance and what database user to connect with.
-    config
-        .setDataSourceClassName("com.microsoft.sqlserver.jdbc.SQLServerDataSource");
-    config.setUsername(DB_USER); // e.g. "root", "sqlserver"
-    config.setPassword(DB_PASS); // e.g. "my-password"
-    config.addDataSourceProperty("databaseName", DB_NAME);
-
-    config.addDataSourceProperty("socketFactoryClass",
-        "com.google.cloud.sql.sqlserver.SocketFactory");
-    config.addDataSourceProperty("socketFactoryConstructorArg", INSTANCE_CONNECTION_NAME);
-
-    // ... Specify additional connection properties here.
-
-    // [START_EXCLUDE]
-
-    // [START cloud_sql_sqlserver_servlet_limit]
-    // maximumPoolSize limits the total number of concurrent connections this pool will keep. Ideal
-    // values for this setting are highly variable on app design, infrastructure, and database.
-    config.setMaximumPoolSize(5);
-    // minimumIdle is the minimum number of idle connections Hikari maintains in the pool.
-    // Additional connections will be established to meet this value unless the pool is full.
-    config.setMinimumIdle(5);
-    // [END cloud_sql_sqlserver_servlet_limit]
-
-    // [START cloud_sql_sqlserver_servlet_timeout]
-    // setConnectionTimeout is the maximum number of milliseconds to wait for a connection checkout.
-    // Any attempt to retrieve a connection from this pool that exceeds the set limit will throw an
-    // SQLException.
-    config.setConnectionTimeout(10000); // 10 seconds
-    // idleTimeout is the maximum amount of time a connection can sit in the pool. Connections that
-    // sit idle for this many milliseconds are retried if minimumIdle is exceeded.
-    config.setIdleTimeout(600000); // 10 minutes
-    // [END cloud_sql_sqlserver_servlet_timeout]
-
-    // [START cloud_sql_sqlserver_servlet_backoff]
-    // Hikari automatically delays between failed connection attempts, eventually reaching a
-    // maximum delay of `connectionTimeout / 2` between attempts.
-    // [END cloud_sql_sqlserver_servlet_backoff]
-
-    // [START cloud_sql_sqlserver_servlet_lifetime]
-    // maxLifetime is the maximum possible lifetime of a connection in the pool. Connections that
-    // live longer than this many milliseconds will be closed and reestablished between uses. This
-    // value should be several minutes shorter than the database's timeout value to avoid unexpected
-    // terminations.
-    config.setMaxLifetime(1800000); // 30 minutes
-    // [END cloud_sql_sqlserver_servlet_lifetime]
-
-    // [END_EXCLUDE]
-
-    // Initialize the connection pool using the configuration object.
-    DataSource pool = new HikariDataSource(config);
-    // [END cloud_sql_sqlserver_servlet_create]
-    return pool;
-  }
-
-  private void createTable(DataSource pool) throws SQLException {
-    // Safely attempt to create the table schema.
-    try (Connection conn = pool.getConnection()) {
-      PreparedStatement createTableStatement = conn.prepareStatement(
-          "IF NOT EXISTS ("
-              + "SELECT * FROM sysobjects WHERE name='votes' and xtype='U')"
-              + "CREATE TABLE votes ("
-              + "vote_id INT NOT NULL IDENTITY,"
-              + "time_cast DATETIME NOT NULL,"
-              + "candidate VARCHAR(6) NOT NULL,"
-              + "PRIMARY KEY (vote_id));"
-      );
-      createTableStatement.execute();
-    }
-  }
-
   @Override
   public void contextDestroyed(ServletContextEvent event) {
     // This function is called when the Servlet is destroyed.
@@ -143,11 +52,16 @@ public class ConnectionPoolContextListener implements ServletContextListener {
     // that can be used to connect to.
     DataSource pool = (DataSource) event.getServletContext().getAttribute("my-pool");
     if (pool == null) {
-      pool = createConnectionPool();
+      if (System.getenv("INSTANCE_HOST") != null) {
+        pool = TcpConnectionPoolFactory.createConnectionPool();
+      } else {
+        pool = ConnectorConnectionPoolFactory.createConnectionPool();
+      }
       event.getServletContext().setAttribute("my-pool", pool);
     }
     try {
-      createTable(pool);
+      // from src/main/java/com/example/cloudsql/Utils.java
+      Utils.createTable(pool);
     } catch (SQLException ex) {
       throw new RuntimeException("Unable to verify table schema. Please double check the steps"
           + "in the README and try again.", ex);
