@@ -19,12 +19,9 @@ package com.example.cloudsql;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Timestamp;
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.servlet.ServletException;
@@ -35,71 +32,17 @@ import javax.servlet.http.HttpServletResponse;
 import javax.sql.DataSource;
 
 
-
 @WebServlet(name = "Index", value = "")
 public class IndexServlet extends HttpServlet {
 
   private static final Logger LOGGER = Logger.getLogger(IndexServlet.class.getName());
 
-  class TemplateData {
-
-    public int tabCount;
-    public int spaceCount;
-    public List<Vote> recentVotes;
-
-    public TemplateData(int tabCount, int spaceCount, List<Vote> recentVotes) {
-      this.tabCount = tabCount;
-      this.spaceCount = spaceCount;
-      this.recentVotes = recentVotes;
-    }
-  }
-
   public TemplateData getTemplateData(DataSource pool) throws ServletException {
-
-    int tabCount;
-    int spaceCount;
-    List<Vote> recentVotes = new ArrayList<>();
-    try (Connection conn = pool.getConnection()) {
-      // PreparedStatements are compiled by the database immediately and executed at a later date.
-      // Most databases cache previously compiled queries, which improves efficiency.
-      PreparedStatement voteStmt = conn.prepareStatement(
-          "SELECT TOP(5) candidate, time_cast FROM votes ORDER BY time_cast DESC");
-      // Execute the statement
-      ResultSet voteResults = voteStmt.executeQuery();
-      // Convert a ResultSet into Vote objects
-      while (voteResults.next()) {
-        String candidate = voteResults.getString(1);
-        Timestamp timeCast = voteResults.getTimestamp(2);
-        Vote vote = new Vote(candidate.trim(), timeCast);
-        recentVotes.add(vote);
-      }
-
-      // PreparedStatements can also be executed multiple times with different arguments. This can
-      // improve efficiency, and project a query from being vulnerable to an SQL injection.
-      PreparedStatement voteCountStmt = conn.prepareStatement(
-          "SELECT COUNT(vote_id) FROM votes WHERE candidate=?");
-
-      voteCountStmt.setString(1, "TABS");
-      ResultSet tabResult = voteCountStmt.executeQuery();
-      tabResult.next(); // Move to the first result
-      tabCount = tabResult.getInt(1);
-
-      voteCountStmt.setString(1, "SPACES");
-      ResultSet spaceResult = voteCountStmt.executeQuery();
-      spaceResult.next(); // Move to the first result
-      spaceCount = spaceResult.getInt(1);
-
+    try {
+      return TemplateData.getTemplateData(pool);
     } catch (SQLException ex) {
-      // If something goes wrong, the application needs to react appropriately. This might mean
-      // getting a new connection and executing the query again, or it might mean redirecting the
-      // user to a different page to let them know something went wrong.
-      throw new ServletException("Unable to successfully connect to the database. Please check the "
-          + "steps in the README and try again.", ex);
+      throw new ServletException(ex);
     }
-
-    TemplateData templateData = new TemplateData(tabCount, spaceCount, recentVotes);
-
-    return templateData;
   }
 
   @Override
@@ -122,12 +65,9 @@ public class IndexServlet extends HttpServlet {
   public void doPost(HttpServletRequest req, HttpServletResponse resp)
       throws IOException {
     // Get the team from the request and record the time of the vote.
-    String team = req.getParameter("team");
-    if (team != null) {
-      team = team.toUpperCase();
-    }
+    String team = Utils.validateTeam(req.getParameter("team"));
     Timestamp now = new Timestamp(new Date().getTime());
-    if (team == null || (!team.equals("TABS") && !team.equals("SPACES"))) {
+    if (team == null) {
       resp.setStatus(400);
       resp.getWriter().append("Invalid team specified.");
       return;
