@@ -29,11 +29,13 @@ import com.google.cloud.compute.v1.NetworkInterface;
 import com.google.cloud.compute.v1.Operation;
 import java.io.IOException;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 public class CreateInstance {
 
   public static void main(String[] args)
-      throws IOException, InterruptedException, ExecutionException {
+      throws IOException, InterruptedException, ExecutionException, TimeoutException {
     // TODO(developer): Replace these variables before running the sample.
     String project = "your-project-id";
     String zone = "zone-name";
@@ -44,15 +46,18 @@ public class CreateInstance {
 
   // Create a new instance with the provided "instanceName" value in the specified project and zone.
   public static void createInstance(String project, String zone, String instanceName)
-      throws IOException, InterruptedException, ExecutionException {
+      throws IOException, InterruptedException, ExecutionException, TimeoutException {
     // Below are sample values that can be replaced.
-    // machineType: machine type of the VM being created. This value uses the format zones/{zone}/machineTypes/{type_name}. For a list of machine types, see https://cloud.google.com/compute/docs/machine-types
-    // sourceImage: path to the operating system image to mount. For details about images you can mount, see https://cloud.google.com/compute/docs/images
+    // machineType: machine type of the VM being created.
+    // *   This value uses the format zones/{zone}/machineTypes/{type_name}.
+    // *   For a list of machine types, see https://cloud.google.com/compute/docs/machine-types
+    // sourceImage: path to the operating system image to mount.
+    // *   For details about images you can mount, see https://cloud.google.com/compute/docs/images
     // diskSizeGb: storage size of the boot disk to attach to the instance.
     // networkName: network interface to associate with the instance.
     String machineType = String.format("zones/%s/machineTypes/n1-standard-1", zone);
     String sourceImage = String
-        .format("projects/debian-cloud/global/images/family/%s", "debian-10");
+        .format("projects/debian-cloud/global/images/family/%s", "debian-11");
     long diskSizeGb = 10L;
     String networkName = "default";
 
@@ -69,12 +74,15 @@ public class CreateInstance {
               .setType(Type.PERSISTENT.toString())
               .setDeviceName("disk-1")
               .setInitializeParams(
-                  AttachedDiskInitializeParams.newBuilder().setSourceImage(sourceImage)
-                      .setDiskSizeGb(diskSizeGb).build())
+                  AttachedDiskInitializeParams.newBuilder()
+                      .setSourceImage(sourceImage)
+                      .setDiskSizeGb(diskSizeGb)
+                      .build())
               .build();
 
       // Use the network interface provided in the networkName argument.
-      NetworkInterface networkInterface = NetworkInterface.newBuilder().setName(networkName)
+      NetworkInterface networkInterface = NetworkInterface.newBuilder()
+          .setName(networkName)
           .build();
 
       // Bind `instanceName`, `machineType`, `disk`, and `networkInterface` to an instance.
@@ -92,13 +100,14 @@ public class CreateInstance {
       InsertInstanceRequest insertInstanceRequest = InsertInstanceRequest.newBuilder()
           .setProject(project)
           .setZone(zone)
-          .setInstanceResource(instanceResource).build();
+          .setInstanceResource(instanceResource)
+          .build();
 
       OperationFuture<Operation, Operation> operation = instancesClient.insertAsync(
           insertInstanceRequest);
 
       // Wait for the operation to complete.
-      Operation response = operation.get();
+      Operation response = operation.get(3, TimeUnit.MINUTES);
 
       if (response.hasError()) {
         System.out.println("Instance creation failed ! ! " + response);
