@@ -25,11 +25,13 @@ import com.google.cloud.compute.v1.Operation.Status;
 import com.google.cloud.compute.v1.StartInstanceRequest;
 import java.io.IOException;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 public class StartInstance {
 
   public static void main(String[] args)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     // TODO(developer): Replace these variables before running the sample.
     /* project: project ID or project number of the Cloud project your instance belongs to.
        zone: name of the zone your instance belongs to.
@@ -43,7 +45,7 @@ public class StartInstance {
 
   // Starts a stopped Google Compute Engine instance (with unencrypted disks).
   public static void startInstance(String project, String zone, String instanceName)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     /* Initialize client that will be used to send requests. This client only needs to be created
        once, and can be reused for multiple requests. After completing all of your requests, call
        the `instancesClient.close()` method on the client to safely
@@ -61,7 +63,7 @@ public class StartInstance {
           startInstanceRequest);
 
       // Wait for the operation to complete.
-      Operation response = operation.get();
+      Operation response = operation.get(3, TimeUnit.MINUTES);
 
       if (response.getStatus() == Status.DONE) {
         System.out.println("Instance started successfully ! ");
