@@ -35,11 +35,14 @@ import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.client.methods.RequestBuilder;
 import org.apache.http.client.utils.URIBuilder;
+import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.HttpClients;
 
 public class FhirResourceSearchGet {
   private static final String FHIR_NAME =
       "projects/%s/locations/%s/datasets/%s/fhirStores/%s/fhir/%s";
+  // The endpoint URL for the Healthcare API. Required for HttpClient.
+  private static final String API_ENDPOINT = "https://healthcare.googleapis.com";
   private static final JsonFactory JSON_FACTORY = new JacksonFactory();
   private static final NetHttpTransport HTTP_TRANSPORT = new NetHttpTransport();
 
@@ -47,21 +50,20 @@ public class FhirResourceSearchGet {
       throws IOException, URISyntaxException {
     // String resourceName =
     //    String.format(
-    //        FHIR_NAME, "project-id", "region-id", "dataset-id", "store-id", "resource-type");
+    //        FHIR_NAME, "project-id", "region-id", "dataset-id", "fhir-store-id");
+    // String resourceType = "Patient";
 
-    // Initialize the client, which will be used to interact with the service.
-    CloudHealthcare client = createClient();
-
+    // Instantiate the client, which will be used to interact with the service.
     HttpClient httpClient = HttpClients.createDefault();
-    String uri = String.format("%sv1/%s", client.getRootUrl(), resourceName);
+    String uri = String.format("%s/v1/%s", API_ENDPOINT, resourceName);
     URIBuilder uriBuilder = new URIBuilder(uri).setParameter("access_token", getAccessToken());
+    // To set additional parameters for search filtering, add them to the URIBuilder. For
+    // example, to search for a Patient with the family name "Smith", specify the following:
+    // uriBuilder.setParameter("family:exact", "Smith");
 
     HttpUriRequest request =
         RequestBuilder.get()
             .setUri(uriBuilder.build())
-            .addHeader("Content-Type", "application/fhir+json")
-            .addHeader("Accept-Charset", "utf-8")
-            .addHeader("Accept", "application/fhir+json; charset=utf-8")
             .build();
 
     // Execute the request and process the results.
@@ -74,7 +76,7 @@ public class FhirResourceSearchGet {
       responseEntity.writeTo(System.err);
       throw new RuntimeException();
     }
-    System.out.println("FHIR resource search results: ");
+    System.out.println("FHIR resource GET search results: ");
     responseEntity.writeTo(System.out);
   }
 
@@ -84,7 +86,6 @@ public class FhirResourceSearchGet {
     GoogleCredentials credential =
         GoogleCredentials.getApplicationDefault()
             .createScoped(Collections.singleton(CloudHealthcareScopes.CLOUD_PLATFORM));
-
     // Create a HttpRequestInitializer, which will provide a baseline configuration to all requests.
     HttpRequestInitializer requestInitializer =
         request -> {
@@ -92,7 +93,6 @@ public class FhirResourceSearchGet {
           request.setConnectTimeout(60000); // 1 minute connect timeout
           request.setReadTimeout(60000); // 1 minute read timeout
         };
-
     // Build the client for interacting with the service.
     return new CloudHealthcare.Builder(HTTP_TRANSPORT, JSON_FACTORY, requestInitializer)
         .setApplicationName("your-application-name")
@@ -103,7 +103,7 @@ public class FhirResourceSearchGet {
     GoogleCredentials credential =
         GoogleCredentials.getApplicationDefault()
             .createScoped(Collections.singleton(CloudHealthcareScopes.CLOUD_PLATFORM));
-    
+
     return credential.refreshAccessToken().getTokenValue();
   }
 }
