@@ -41,6 +41,8 @@ import org.apache.http.impl.client.HttpClients;
 public class FhirResourceSearchPost {
   private static final String FHIR_NAME =
       "projects/%s/locations/%s/datasets/%s/fhirStores/%s/fhir/%s";
+  // The endpoint URL for the Healthcare API. Required for HttpClient.
+  private static final String API_ENDPOINT = "https://healthcare.googleapis.com";
   private static final JsonFactory JSON_FACTORY = new JacksonFactory();
   private static final NetHttpTransport HTTP_TRANSPORT = new NetHttpTransport();
 
@@ -50,12 +52,15 @@ public class FhirResourceSearchPost {
     //    String.format(
     //        FHIR_NAME, "project-id", "region-id", "dataset-id", "store-id", "resource-type");
 
-    // Initialize the client, which will be used to interact with the service.
-    CloudHealthcare client = createClient();
-
+    // Instantiate the client, which will be used to interact with the service.
     HttpClient httpClient = HttpClients.createDefault();
-    String uri = String.format("%sv1/%s/_search", client.getRootUrl(), resourceName);
+    String uri = String.format("%s/v1/%s/_search", API_ENDPOINT, resourceName);
     URIBuilder uriBuilder = new URIBuilder(uri).setParameter("access_token", getAccessToken());
+    // To set additional parameters for search filtering, add them to the URIBuilder. For
+    // example, to search for a Patient with the family name "Smith", specify the following:
+    // uriBuilder.setParameter("family:exact", "Smith");
+
+    // Set a body otherwise HttpClient complains there is no Content-Length set.
     StringEntity requestEntity = new StringEntity("");
 
     HttpUriRequest request =
@@ -78,7 +83,7 @@ public class FhirResourceSearchPost {
       responseEntity.writeTo(System.err);
       throw new RuntimeException();
     }
-    System.out.println("FHIR resource search results: ");
+    System.out.println("FHIR resource POST search results: ");
     responseEntity.writeTo(System.out);
   }
 
@@ -107,7 +112,7 @@ public class FhirResourceSearchPost {
     GoogleCredentials credential =
         GoogleCredentials.getApplicationDefault()
             .createScoped(Collections.singleton(CloudHealthcareScopes.CLOUD_PLATFORM));
-    
+
     return credential.refreshAccessToken().getTokenValue();
   }
 }
