@@ -146,7 +146,7 @@ public class FhirResourceTests {
     FhirResourceSearchGet.fhirResourceSearchGet(resourcePath);
 
     String output = bout.toString();
-    assertThat(output, containsString("FHIR resource search results:"));
+    assertThat(output, containsString("FHIR resource GET search results:"));
   }
 
   @Test
@@ -154,7 +154,7 @@ public class FhirResourceTests {
     FhirResourceSearchPost.fhirResourceSearchPost(resourcePath);
 
     String output = bout.toString();
-    assertThat(output, containsString("FHIR resource search results:"));
+    assertThat(output, containsString("FHIR resource POST search results:"));
   }
 
   @Test
