@@ -65,14 +65,34 @@ public class UptimeIT {
 
   @Test
   public void test2_UpdateUptimeCheck() throws Exception {
-    UptimeSample.main("update", "-n", config.getDisplayName(), "-a", "/updated");
-    assertThat(bout.toString()).contains("/updated");
+    int maxAttempts = 5;
+    for (int count = 0; count < maxAttempts; count++) {
+      try {
+        UptimeSample.main("update", "-n", config.getDisplayName(), "-a", "/updated");
+        assertThat(bout.toString()).contains("/updated");
+        return;
+      } catch (Exception e) {
+        System.out.println("Attempt %d failed. Retrying. Error: " + e.toString());
+        Thread.sleep(3000);
+      }
+    }
+    assertThat(false);
   }
 
   @Test
   public void test2_GetUptimeCheck() throws Exception {
-    UptimeSample.main("get", "-n", config.getDisplayName());
-    assertThat(bout.toString()).contains(config.getDisplayName());
+    int maxAttempts = 5;
+    for (int count = 0; count < maxAttempts; count++) {
+      try {
+        UptimeSample.main("get", "-n", config.getDisplayName());
+        assertThat(bout.toString()).contains(config.getDisplayName());
+        return;
+      } catch (Exception e) {
+        System.out.println("Attempt %d failed. Retrying. Error: " + e.toString());
+        Thread.sleep(3000);
+      }
+    }
+    assertThat(false);
   }
 
   @Test
@@ -94,6 +114,16 @@ public class UptimeIT {
 
   @Test
   public void test3_DeleteUptimeCheck() throws Exception {
-    UptimeSample.main("delete", "-n", config.getDisplayName());
+    int maxAttempts = 5;
+    for (int count = 0; count < maxAttempts; count++) {
+      try {
+        UptimeSample.main("delete", "-n", config.getDisplayName());
+        return;
+      } catch (Exception e) {
+        System.out.println("Attempt " + count + " failed. Retrying. Error: " + e.toString());
+        Thread.sleep(3000);
+      }
+    }
+    assertThat(false);
   }
 }
