@@ -14,28 +14,39 @@
  * limitations under the License.
  */
 
+// [START cloudrun_jobs_quickstart]
+
 package com.example;
 
 abstract class JobsExample {
-
+  // [START cloudrun_jobs_env_vars]
   // These values are provided automatically by the Cloud Run Jobs runtime.
-  private static String TASK_NUM = System.getenv().getOrDefault("TASK_NUM", "0");
-  private static String ATTEMPT_NUM = System.getenv().getOrDefault("ATTEMPT_NUM", "0");
+  private static String CLOUD_RUN_TASK_INDEX =
+      System.getenv().getOrDefault("CLOUD_RUN_TASK_INDEX", "0");
+  private static String CLOUD_RUN_TASK_ATTEMPT =
+      System.getenv().getOrDefault("CLOUD_RUN_TASK_ATTEMPT", "0");
 
   // User-provided environment variables
   private static int SLEEP_MS = Integer.parseInt(System.getenv().getOrDefault("SLEEP_MS", "0"));
   private static float FAIL_RATE =
       Float.parseFloat(System.getenv().getOrDefault("FAIL_RATE", "0.0"));
+  // [END cloudrun_jobs_env_vars]
 
   // Start script
   public static void main(String[] args) {
-    System.out.println(String.format("Starting Task #%s, Attempt #%s...", TASK_NUM, ATTEMPT_NUM));
+    System.out.println(
+        String.format(
+            "Starting Task #%s, Attempt #%s...", CLOUD_RUN_TASK_INDEX, CLOUD_RUN_TASK_ATTEMPT));
     try {
       runTask(SLEEP_MS, FAIL_RATE);
     } catch (RuntimeException | InterruptedException e) {
-      System.err.println(String.format("Task #%s, Attempt #%s failed.", TASK_NUM, ATTEMPT_NUM));
+      System.err.println(
+          String.format(
+              "Task #%s, Attempt #%s failed.", CLOUD_RUN_TASK_INDEX, CLOUD_RUN_TASK_ATTEMPT));
+      // [START cloudrun_jobs_exit_process]
       // Catch error and denote process-level failure to retry Task
       System.exit(1);
+      // [END cloudrun_jobs_exit_process]
     }
   }
 
@@ -56,6 +67,7 @@ abstract class JobsExample {
     if (Math.random() < failureRate) {
       throw new RuntimeException("Task Failed.");
     }
-    System.out.println(String.format("Completed Task #%s", TASK_NUM));
+    System.out.println(String.format("Completed Task #%s", CLOUD_RUN_TASK_INDEX));
   }
 }
+// [END cloudrun_jobs_quickstart]
